// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SOCIETY: 'SOCIETY',
  OWNER: 'OWNER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const GenderType = {
  PUTRA: 'PUTRA',
  PUTRI: 'PUTRI',
  CAMPUR: 'CAMPUR'
} as const

export type GenderType = (typeof GenderType)[keyof typeof GenderType]


export const BookingStatus = {
  PENDING: 'PENDING',
  DITERIMA: 'DITERIMA',
  DITOLAK: 'DITOLAK'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]